requirements.txt
astroid==2.2.5
certifi==2019.3.9
chardet==3.0.4
colorama==0.4.1
Django==2.2.1
django-ckeditor==5.7.1
django-hitcount==1.3.0
django-js-asset==1.2.2
django-taggit==1.1.0
easy-thumbnails==2.6
funcsigs==1.0.2
idna==2.8
isort==4.3.20
lazy-object-proxy==1.4.1
mccabe==0.6.1
mock==3.0.5
pbr==5.2.0
Pillow==6.0.0
pylint==2.3.1
python-decouple==3.1
pytz==2019.1
requests==2.22.0
simplejson==3.16.0
six==1.12.0
sqlparse==0.3.0
typed-ast==1.3.5
Unidecode==1.0.23
urllib3==1.25.3
wget==3.2
wrapt==1.11.1
